+++ lib/views/widgets/app_input_fields/app_file_picker.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:reusable_editor/view_models/field_cubits/field_cubit/field_cubit.dart';
+import 'package:reusable_editor/view_models/generic_cubits/field_cubit/field_cubit.dart';
 
 // ✅ File Picker
 class AppFilePicker extends StatelessWidget {