+++ lib/views/widgets/app_input_fields/app_switch.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:reusable_editor/styles/app_form_text_styles.dart';
-import 'package:reusable_editor/view_models/field_cubits/field_cubit/field_cubit.dart';
+import 'package:reusable_editor/view_models/generic_cubits/field_cubit/field_cubit.dart';
 
 // ✅ Switch
 class AppSwitch extends StatelessWidget {