+++ lib/view_models/field_cubits/image_field_cubit/image_field_cubit.dart
@@ -21,7 +21,7 @@ class ImageFieldCubit extends Cubit<ImageFieldState> {
     }
 
     debugPrint(
-      'ImageFieldCubit | selectImage | path: ${pickedFile?.path} | error: $error',
+      'ImageFieldCubit | selectImage | path: ${pickedFile?.path} | error: $error state pickedFile: ${state.pickedFile?.name}',
     );
   }